requirements.txt
gitdb2==2.0.6
thingspeak==0.4.1
gTTS==2.0.3
feedparser==5.2.1
gitpython==2.1.11
fabric3==1.14.post1
youtube_dl==2019.7.16
spotipy==2.4.4
FFmpeg==1.4
websockets==7.0
aiohttp==3.5.4
python-dateutil==2.8.0
selenium==3.141.0
paho-mqtt==1.4.0
pexpect==4.7.0
py-enigma==0.1
discord.py==0.16.12
SQLAlchemy==1.2.15
joblib==0.11
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
opencv-python==3.4.4.19
pushbullet.py==0.11.0
twython==3.7.0
mcstatus==2.3.0
numpy==1.12.1